+++ test/AutoReview/TestCodeTest.php
@@ -1,0 +1,39 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Copyright (c) 2018 Andreas Möller.
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE file that was distributed with this source code.
+ *
+ * @see https://github.com/localheinz/json-normalizer
+ */
+
+namespace Localheinz\Json\Normalizer\Test\AutoReview;
+
+use Localheinz\Test\Util\Helper;
+use PHPUnit\Framework;
+
+/**
+ * @internal
+ */
+final class TestCodeTest extends Framework\TestCase
+{
+    use Helper;
+
+    public function testProductionClassesAreAbstractOrFinal(): void
+    {
+        $this->assertClassesAreAbstractOrFinal(__DIR__ . '/../../src');
+    }
+
+    public function testProductionClassesHaveTests(): void
+    {
+        $this->assertClassesHaveTests(
+            __DIR__ . '/../../src',
+            'Localheinz\\Json\\Normalizer\\',
+            'Localheinz\\Json\\Normalizer\\Test\\Unit\\'
+        );
+    }
+}
